requirements.txt
openai~=0.27.8
requests~=2.31.0
telebot~=0.0.5
python-dotenv~=1.0.0